Safari-609.1.11.1/LayoutTests/fast/canvas/webgl/webgl2-texStorage-expected.txt

CONSOLE MESSAGE: line 58: WebGL: INVALID_OPERATION: texStorage2D: texStorage2D already called on this texture
CONSOLE MESSAGE: line 61: WebGL: INVALID_OPERATION: texImage2D: texStorage() called on this texture previously
Test that texStorage2D() works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
asdf
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
PASS gl.getError() is gl.NO_ERROR
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS gl.getError() is not gl.NO_ERROR
PASS gl.getError() is not gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
